Revert "fix: reject sibling-prefix path escapes in all extractors (#157)"
This reverts commit 2d593e429e4335a440d5ec0a0daad99f639b72b2.
